Add analysis filter combination base classes
Analysis filter base classed create a hybrid custom analysis type. The
main goal would be data analysis through a single user interface,
but it additionally implements the required functionality for filtering
the input ions by custom criteria and allowing further analysis.

Underlying node is actually a data filter with analysis view support,
not analysis node with data filter support. This is due to how AP Suite
function. As a result, the GetIonData methods get the output of the
node, not the input parent ion data as analysis nodes typically would.
Ensure that all analysis portions of the hybrid node use GetOwnerIonData
instead of GetIonData.

Publicly expose some core node properties. Should not be a breaking
change as the properties were not virtual.