Add helper for setting read-only deferred member
RequiredSetOnce wraps a generic value and throws if either the
wrapped value is accessed before set (required) or has multiple
attempts to set the value (set once).

This is useful for deferred initialization of values that conceptually
should be read-only but need to be set after the constructor. A common
case will be having the IoC container resolve a type in a factory
class but then have the factory wrapper set a unique instance ID or
other server.