Remove properties for allowing overriding ranges
The problem turns out to not be so much about the recursive definition
issue where the definition of the filter could depend on the IonTypes
section being read, causing a cyclical definition issue that is
incoherent. This attempt was to make a default false property to allow
a developer to allow this with the onus on them to ensure consistency.

However the real issue is simply the data invalidation an update issue.
On basically any other data change, the filter is invalidated, causing
overridden ranges to potentially require a change. AP Suite solves this
problem by locking ROIs with overridden ranges, so it is tied to an
unchanging spatial region. This can't be done with filters. The solution
here instead was to implement a new AP Suite native node to extract and
store indices of all filtered ion independent of the filter code, thus
effectively locking the region to allow overriding.