Revert changes to Resources RangeManager
Using the owner node as the source for RangeManager in Resources
was a fix for incorrect resolution in AP Suite. This is and has been
more appropriately fixed in AP Suite, so this adjustment is no longer
appropriate. It is more consistent to have the range manager use
the current node ID and have AP Suite resolve accordingly rather than
expect every extension to know to use the owner instead.