Add common controls
AppendingTextBox box is a control that can simulate a text console
output by consistently displaying a string collection that can scroll to
follow output. Previously defined only for Python library, but moved
here to be available to any extension.

ButtonOverlayControl is being modified to accept an optional Cancel
command that replaced the disabled update button in the overlay for
long running tasks. Visibilty of the overlay is better handled to only
remove after update command is finished, not immediately.
Clean up template creation to use relative binding sources