Add helper to simplify child node creation
If an extension wants to create other nodes, it's very likely that the
next step is very likely to be to interact with that new node. That
requires resolving the new node ID.IonDataExtensions
This CreateChildNode method registers a temporary event handler
to capture the created node event that broadcasts the new ID and
returns it from the method. This allows creating a new node and
immediately returning the new ID.