Fix segmented child filter issues
Registration of SegmentedRoiNode must be scoped to each utilizing
extension, else the same constant name for all registrations will
collide in dependency resolution.

Display a read only Filter Value in properties of each child node to
help identify the underlying value, which can be useful if names are
ambiguous, or there is a desire to match up the filter to the data
section visualizer.